+++ java/src/main/java/io/pingui/CliRunMode.java
@@ -5,6 +5,7 @@ public enum CliRunMode {
     GUI,
     EXPORT,
     TELEMETRY_RETENTION,
+    TELEMETRY_DUMP,
     DAEMON,
     STOP,
     STATUS